+++ packages/app-degree-pages/src/core/services/degree-data-manager-service.js
@@ -46,10 +46,26 @@ function filterData({
     !isAccelConcValid(acceleratedConcurrent) ||
     row[acceleratedConcurrent.value]?.length > 0;
 
-  const filterByKeyword = (resolver, searchTerm) => {
+  const filterByKeywordTitle = (resolver, searchTerm) => {
     if (!searchTerm) return true;
     const regex = new RegExp(searchTerm, "i");
-    return regex.test(resolver.getFullDescription());
+    const title = regex.test(resolver.getMajorDesc());
+    if (title) {
+      return true;
+    }
+
+    return false;
+  };
+
+  const filterByKeywordDescription = (resolver, searchTerm) => {
+    if (!searchTerm) return true;
+    const regex = new RegExp(searchTerm, "i");
+    const description = regex.test(resolver.getFullDescription());
+    if (description) {
+      return true;
+    }
+
+    return false;
   };
 
   const filterByBlacklist = resolver =>
@@ -81,13 +97,39 @@ function filterData({
       filterByDepartmentCode(resolver) &&
       filterByCampus(resolver) &&
       filterByAcceleratedConcurrent(row) &&
-      filterByKeyword(resolver, keyword) &&
       filterByBlacklist(resolver) &&
       filterGraduateCerts(resolver)
     );
   };
 
-  return programs.filter(applyFilters);
+  const filteredByTitle = [];
+  const filteredByDescription = [];
+
+  const newPrograms = programs.filter(degree => {
+    const resolver = degreeDataPropResolverService(degree);
+
+    const isFilteredByTitle = filterByKeywordTitle(resolver, keyword);
+    const isFilteredByDescription = filterByKeywordDescription(
+      resolver,
+      keyword
+    );
+
+    if (
+      (isFilteredByTitle || isFilteredByDescription) &&
+      applyFilters(degree)
+    ) {
+      if (isFilteredByTitle) {
+        filteredByTitle.push(degree);
+      } else {
+        filteredByDescription.push(degree);
+      }
+      return true;
+    }
+
+    return false;
+  });
+
+  return filteredByTitle.concat(filteredByDescription);
 }
 
 /**