+++ packages/app-degree-pages/CHANGELOG.md
@@ -1,3 +1,12 @@
+# [@asu/app-degree-pages-v3.0.3](https://github.com/asu/asu-unity-stack/compare/@asu/app-degree-pages-v3.0.2...@asu/app-degree-pages-v3.0.3) (2025-04-08)
+
+
+### Bug Fixes
+
+* **app-degree-pages:** changes ([43046a8](https://github.com/asu/asu-unity-stack/commit/43046a8f36984a5cfaf583051697ac546291ca50))
+* **app-degree-pages:** improve the search relevancy by displaying title matches first ([ad9b0bc](https://github.com/asu/asu-unity-stack/commit/ad9b0bc4f0f9373905e72f22f46697fe29be656c))
+* **app-degree-pages:** improving search relevancy ([20e4e20](https://github.com/asu/asu-unity-stack/commit/20e4e20f31c609be24cd8a4f97957c797c2b38ce))
+
 # [@asu/app-degree-pages-v3.0.2](https://github.com/asu/asu-unity-stack/compare/@asu/app-degree-pages-v3.0.1...@asu/app-degree-pages-v3.0.2) (2025-03-03)