+++ packages/app-rfi/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@asu/app-rfi-v3.3.0](https://github.com/asu/asu-unity-stack/compare/@asu/app-rfi-v3.2.0...@asu/app-rfi-v3.3.0) (2025-04-01)
+
+
+### Features
+
+* **app-rfi:** simple change to rebuild release ([13f8a3b](https://github.com/asu/asu-unity-stack/commit/13f8a3b546f4f29d47e3c895e9f48b6058415744))
+
 # [@asu/app-rfi-v3.2.0](https://github.com/asu/asu-unity-stack/compare/@asu/app-rfi-v3.1.0...@asu/app-rfi-v3.2.0) (2025-03-05)